# Use rand.Perm to shuffle data