The idea is the same as in honeycomb's MergeDestructive - shuffle the data in-place without creating temporal summary. This version uses `rand.Perm` to achieve that.

```
benchmark                                         old ns/op     new ns/op     delta
BenchmarkTDigestMerge/compression=1_n=1-4         3003          3454          +15.02%
BenchmarkTDigestMerge/compression=1_n=10-4        22186         17029         -23.24%
BenchmarkTDigestMerge/compression=1_n=100-4       192510        156827        -18.54%
BenchmarkTDigestMerge/compression=10_n=1-4        27901         19703         -29.38%
BenchmarkTDigestMerge/compression=10_n=10-4       182637        142998        -21.70%
BenchmarkTDigestMerge/compression=10_n=100-4      1609905       1389554       -13.69%
BenchmarkTDigestMerge/compression=20_n=1-4        64749         45925         -29.07%
BenchmarkTDigestMerge/compression=20_n=10-4       394181        311002        -21.10%
BenchmarkTDigestMerge/compression=20_n=100-4      3527811       3053645       -13.44%
BenchmarkTDigestMerge/compression=30_n=1-4        102763        78263         -23.84%
BenchmarkTDigestMerge/compression=30_n=10-4       613174        561055        -8.50%
BenchmarkTDigestMerge/compression=30_n=100-4      5467320       5042737       -7.77%
BenchmarkTDigestMerge/compression=50_n=1-4        191938        152864        -20.36%
BenchmarkTDigestMerge/compression=50_n=10-4       1125494       1026854       -8.76%
BenchmarkTDigestMerge/compression=50_n=100-4      10034623      9390289       -6.42%
BenchmarkTDigestMerge/compression=100_n=1-4       445269        416063        -6.56%
BenchmarkTDigestMerge/compression=100_n=10-4      2727377       2578484       -5.46%
BenchmarkTDigestMerge/compression=100_n=100-4     25377255      23337689      -8.04%

benchmark                                         old allocs     new allocs     delta
BenchmarkTDigestMerge/compression=1_n=1-4         10             10             +0.00%
BenchmarkTDigestMerge/compression=1_n=10-4        29             20             -31.03%
BenchmarkTDigestMerge/compression=1_n=100-4       209            116            -44.50%
BenchmarkTDigestMerge/compression=10_n=1-4        10             10             +0.00%
BenchmarkTDigestMerge/compression=10_n=10-4       28             19             -32.14%
BenchmarkTDigestMerge/compression=10_n=100-4      210            111            -47.14%
BenchmarkTDigestMerge/compression=20_n=1-4        10             10             +0.00%
BenchmarkTDigestMerge/compression=20_n=10-4       28             19             -32.14%
BenchmarkTDigestMerge/compression=20_n=100-4      210            111            -47.14%
BenchmarkTDigestMerge/compression=30_n=1-4        10             10             +0.00%
BenchmarkTDigestMerge/compression=30_n=10-4       28             19             -32.14%
BenchmarkTDigestMerge/compression=30_n=100-4      210            111            -47.14%
BenchmarkTDigestMerge/compression=50_n=1-4        10             10             +0.00%
BenchmarkTDigestMerge/compression=50_n=10-4       28             19             -32.14%
BenchmarkTDigestMerge/compression=50_n=100-4      210            111            -47.14%
BenchmarkTDigestMerge/compression=100_n=1-4       10             10             +0.00%
BenchmarkTDigestMerge/compression=100_n=10-4      28             19             -32.14%
BenchmarkTDigestMerge/compression=100_n=100-4     210            111            -47.14%

benchmark                                         old bytes     new bytes     delta
BenchmarkTDigestMerge/compression=1_n=1-4         495           596           +20.40%
BenchmarkTDigestMerge/compression=1_n=10-4        1772          1469          -17.10%
BenchmarkTDigestMerge/compression=1_n=100-4       11872         8695          -26.76%
BenchmarkTDigestMerge/compression=10_n=1-4        2776          3623          +30.51%
BenchmarkTDigestMerge/compression=10_n=10-4       9682          8196          -15.35%
BenchmarkTDigestMerge/compression=10_n=100-4      76792         54860         -28.56%
BenchmarkTDigestMerge/compression=20_n=1-4        5536          7209          +30.22%
BenchmarkTDigestMerge/compression=20_n=10-4       19166         16160         -15.68%
BenchmarkTDigestMerge/compression=20_n=100-4      153729        108539        -29.40%
BenchmarkTDigestMerge/compression=30_n=1-4        7970          10657         +33.71%
BenchmarkTDigestMerge/compression=30_n=10-4       28128         24432         -13.14%
BenchmarkTDigestMerge/compression=30_n=100-4      226685        159116        -29.81%
BenchmarkTDigestMerge/compression=50_n=1-4        12783         16748         +31.02%
BenchmarkTDigestMerge/compression=50_n=10-4       46111         39512         -14.31%
BenchmarkTDigestMerge/compression=50_n=100-4      373601        264224        -29.28%
BenchmarkTDigestMerge/compression=100_n=1-4       25697         33696         +31.13%
BenchmarkTDigestMerge/compression=100_n=10-4      94255         79988         -15.14%
BenchmarkTDigestMerge/compression=100_n=100-4     768160        545184        -29.03%
```