@vmihailenco 2018-11-08
I also added `MergeDestructive` from honeycomb fork which leaves summary in unsorted state and therefore is faster than `Merge` if you are doing thousands of merges.