@caio 2018-11-08
Looks great! Thanks!!

RE `MergeDestructive`: thinking a bit about it, since it leaves the other digest in a completely useless state (I'm fine with it, name and docs should be enough to prevent misuse) I wonder if it would be worth to invest time in a `MergeFromBuffer` (or whatever name...) directly: no need to construct a tdigest at all and if we chage the payload to not use delta nor varints the whole merge logic can be done straight from the byte buffer.